// src/exceptions/error-code.enum.ts
export enum ErrorCodeEnum {
    UNAUTHORIZED = 600,
    INVALID_CREDENTIALS = 601,
    INVALID_USER = 602,
    NEW_PASSWORD_NOT_CONFIRMED = 603,
    OLD_PASSWORD_AND_NEW_ARE_MATCHED = 604,
    INCORRECT_PASSWORD = 605,
    USERNAME_OR_EMAIL_ALREADY_EXIST = 606,
    INVALID_TOKEN = 607,
    ALLOWED_FOR_ADMIN_ONLY = 608,
}